Altitude and Speed
Added public variables for altitude and speed for the player to see. "Helpful" variables and earth_rad are now in kilometers. Renamed update_axes() to update_axis_variables() and added the update_trajectory_variables() function. Everything can be updated in real time from unity (this needs to be changed in the game for efficiency).