Added control parameters
Added parameters allow for the user to control the satellite. Added functions to compute the velocity and deltaV budget. Begun work on adding deltaV to the satellite (in progress). Rearranged how pos is calculated. Changed new_orbit_2D() to compute_position_vec() and moved the E computations out into Update().